Rebase into: Refactor each subcommand to its own file
